+++ components/hypeauditor/actions/get-tiktok-report/get-tiktok-report.mjs
@@ -1,0 +1,31 @@
+import app from "../../hypeauditor.app.mjs";
+
+export default {
+  key: "hypeauditor-get-tiktok-report",
+  name: "Get Tiktok Report",
+  description: "Returns a report about the specified Tiktok channel. [See the documentation](https://hypeauditor.readme.io/reference/get_report_tiktok)",
+  version: "0.0.2",
+  type: "action",
+  props: {
+    app,
+    channelUsername: {
+      propDefinition: [
+        app,
+        "channelUsername",
+      ],
+      optional: false,
+    },
+  },
+
+  async run({ $ }) {
+    const response = await this.app.getTiktokReport({
+      $,
+      data: {
+        channel: this.channelUsername,
+      },
+    });
+    $.export("$summary", `Successfully sent the request. Report state: '${response.result.report_state}'`);
+
+    return response;
+  },
+};