+++ components/openai/actions/list-run-steps/list-run-steps.mjs
@@ -4,7 +4,7 @@ export default {
   key: "openai-list-run-steps",
   name: "List Run Steps (Assistants)",
   description: "Returns a list of run steps belonging to a run. [See the documentation](https://platform.openai.com/docs/api-reference/runs/list-run-steps)",
-  version: "0.0.12",
+  version: "0.0.13",
   type: "action",
   props: {
     openai,